Add new layout components for dashboard sections and remove deprecated pages
