+++ apps/nextjs/src/app/(dashboard)/dashboard/(body)/@center/(.)organizations/page.tsx
@@ -1,0 +1,5 @@
+export default function Center () {
+  return (
+    <h1>center</h1>
+  )
+}
