+++ apps/nextjs/src/app/(dashboard)/dashboard/(body)/@center/default.tsx
@@ -1,0 +1,3 @@
+export default function Default () {
+  return null;
+}
