--- apps/nextjs/src/app/(dashboard)/dashboard/(body)/admin/@right/page.tsx
@@ -1,3 +1,0 @@
-export default function OrganizationsRight() {
-  return <h1> Profiles </h1>;
-}
