+++ apps/nextjs/src/app/(dashboard)/dashboard/(body)/layout.tsx
@@ -1,17 +1,32 @@
-interface DashboardBodyLayoutProps {
+import { DashboardContentBody } from "~/app/(dashboard)/dashboard/_components/content/dashboard-content-body";
+import { DashboardContentSection } from "~/app/(dashboard)/dashboard/_components/content/dashboard-content-top-section";
+
+interface LayoutProps {
   children: React.ReactNode;
+  left: React.ReactNode;
+  center: React.ReactNode;
+  right: React.ReactNode;
 }
 
-export default function DashboardBodyLayout({
+export default function SharedLayout ({
   children,
-}: DashboardBodyLayoutProps) {
+  left,
+  center,
+  right,
+}: LayoutProps) {
   return (
     <section
       role="status"
       aria-label="Loading dashboard content"
       className="flex flex-1 flex-col gap-4 p-4"
     >
-      {children}
+      <div className="grid auto-rows-min gap-4 md:grid-cols-3">
+        <DashboardContentSection>{ left }</DashboardContentSection>
+        <DashboardContentSection>{ center }</DashboardContentSection>
+        <DashboardContentSection>{ right }</DashboardContentSection>
+      </div>
+
+      <DashboardContentBody>{ children }</DashboardContentBody>
     </section>
   );
 }
