--- apps/nextjs/src/app/(dashboard)/dashboard/(body)/pokemon/@left/page.tsx
@@ -1,3 +1,0 @@
-export default function OrganizationsLeft() {
-  return <h1>Organizations Left</h1>;
-}
