--- apps/nextjs/src/app/(dashboard)/dashboard/(body)/pokemon/@right/page.tsx
@@ -1,3 +1,0 @@
-export default function OrganizationsRight() {
-  return <h1>Organizations Right</h1>;
-}
