--- apps/nextjs/src/app/(dashboard)/dashboard/(body)/settings/layout.tsx
@@ -1,3 +1,0 @@
-import SharedLayout from "~/app/(dashboard)/dashboard/_components/layout-shared";
-
-export default SharedLayout;
