--- apps/nextjs/src/app/(dashboard)/dashboard/(body)/tournaments/@center/page.tsx
@@ -1,3 +1,0 @@
-export default function OrganizationsCenter() {
-  return <h1>Organizations Center</h1>;
-}
